requirements.txt
google-cloud-translate==1.7.0
